# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Light novel API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/novel/:novelID": {
            "post": {
                "description": "If the novel is private, the user need to be logged in with the author account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "Get the novel with provided novel id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Novel ID",
                        "name": "NovelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User's Session",
                        "name": "sessionString",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.IncludeSessionString"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.NovelView"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "novel"
                ],
                "summary": "Delete the novel and all the related stuff like volumes, chapters, comments, images with the provided novel id",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Novel ID",
                        "name": "NovelID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "description": "Possible error code: MissingField, InvalidLanguageFormat, TitleTooLong, TaglineTooLong",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "Update the novel metadata with the provided metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Novel ID",
                        "name": "NovelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novel details",
                        "name": "NovelDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NovelMetadata"
                        }
                    },
                    {
                        "description": "User's Session",
                        "name": "sessionString",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IncludeSessionString"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.ErrorJSON"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/novel/create": {
            "post": {
                "description": "Possible error code: MissingField, InvalidLanguageFormat, TitleTooLong, TaglineTooLong",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "Create a new novel with the provided metadata, return the created novel id",
                "parameters": [
                    {
                        "description": "Novel details",
                        "name": "NovelDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NovelMetadata"
                        }
                    },
                    {
                        "description": "User's Session",
                        "name": "sessionString",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IncludeSessionString"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/route.createNovelResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.ErrorJSON"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/novel/from/:userID": {
            "post": {
                "description": "If the user is not logged in, only the public novels will be returned",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "Get all the novels from the user with the provided user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "UserID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User's Session",
                        "name": "sessionString",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IncludeSessionString"
                        }
                    },
                    {
                        "type": "boolean",
                        "name": "adult",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fromDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at",
                            "views",
                            "title"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "OrderByCreatedAt",
                            "OrderByUpdateAt",
                            "OrderByViews",
                            "OrderByTitle"
                        ],
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "SortOrderAsc",
                            "SortOrderDesc"
                        ],
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "x-enum-varnames": [
                            "StatusOngoing",
                            "StatusCompleted",
                            "StatusDropped"
                        ],
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "name": "tagExclude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "toDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.NovelMetadataSmall"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.IncludeSessionString": {
            "type": "object",
            "properties": {
                "session": {
                    "type": "string"
                }
            }
        },
        "model.NovelMetadata": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.NovelStatusID"
                },
                "tagline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "$ref": "#/definitions/model.VisibilityID"
                }
            }
        },
        "model.NovelMetadataSmall": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "boolean"
                },
                "author": {
                    "$ref": "#/definitions/model.UserMetadataSmall"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "rateCount": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalRating": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "model.NovelStatusID": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "StatusOngoing",
                "StatusCompleted",
                "StatusDropped"
            ]
        },
        "model.NovelView": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "boolean"
                },
                "author": {
                    "$ref": "#/definitions/model.UserMetadataSmall"
                },
                "clicks": {
                    "type": "integer"
                },
                "createAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "followCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "rateCount": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TagView"
                    }
                },
                "title": {
                    "type": "string"
                },
                "totalRating": {
                    "type": "integer"
                },
                "updateAt": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "string"
                },
                "volumes": {
                    "type": "integer"
                }
            }
        },
...
        "model.TagView": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.UserMetadataSmall": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.VisibilityID": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "VisibilityPrivate",
                "VisibilityPublic"
            ]
        },
        "route.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12
            ],
            "x-enum-varnames": [
                "BadInput",
                "UserNotFound",
                "WrongPassword",
                "BadPassword",
                "BadUsername",
                "BadDeviceName",
                "BadDisplayname",
                "BadEmail",
                "UserAlreadyExists",
                "InvalidLanguageFormat",
                "TitleTooLong",
                "TaglineTooLong",
                "DescriptionTooLong"
            ]
        },
        "route.ErrorJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/route.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "route.createNovelResult": {
            "type": "object",
            "properties": {
                "novel_id": {
                    "type": "string"
                }
            }
        },
...
    }
}